requirements.txt
Flask
python-dotenv
groq
google-generativeai
pandas
python-docx